Initial commit for adding the System Simulator(SS) in OAI
The feature is to enable a new mode in OAI to run as a System Simulator in
the TTCN env setup. The SS mode runs the eNB in L2-simulator mode and it
needs the Multi UE Proxy running for the functionality to work.

Following is summary details of the changes that are submitted as part of
this commit:

- Added the system simulator functionality as per 36.523 under openair3/SS
 directory

- The SS has the Sidl(Sequans Interface Descriptive Language) interface
towards northbound.

- Following port support at SS are implemented
   -SYS_PORT: For handling SYTEM_CTRL message from TTCN
   -SRB_PORT: For handling RRC_PDU message from TTCN
   -VNG_PORT: For handling Noise generation cmd from TTCN

- The Sidl directory has the structures defintion for the northbound
  interface

- Added the inferace to Multi UE proxy using the UDP interface, the udp
  task module of OAI is re-introduced to manage this interface

- Modified the eNB configuraiton files to add the SS mode configuration
  details

- Added new ITTI messages for SS mode operation

- Added new SS Mode context structures in ran_context.